+++ cli/actions/actions.go
@@ -27,6 +27,8 @@ type SedgeActions interface {
 	SetupContainers(SetupContainersOptions) error
 	RunContainers(RunContainersOptions) error
 	InstallDependencies(InstallDependenciesOptions) error
+	Generate(GenerateOptions) error
+	CreateJWTSecrets(CreateJWTSecretOptions) (string, error)
 }
 
 type sedgeActions struct {